Eliminación componentes modales innecesarios y actualización página edituserprofile para implementar el nuevo modal en el cambio de contraseña
