--- src/Components/ChangePassForm/ChangePassForm.css
@@ -1,5 +1,0 @@
-.ChangePss-wrap {
-  padding: 50px;
-  /* display: flex;
-    flex-direction: row; */
-}
