+++ src/Components/Modal/Modal.css
@@ -16,9 +16,9 @@
   color: rgb(51, 51, 51);
   display: inline;
   /* display: flex; */
-  height: 300px;
+  height: 70%;
   width: 50%;
-  /* max-height: 620px; */
+  max-height: 620px;
   max-width: 500px;
   outline: none;
   padding: 10px;
@@ -33,6 +33,10 @@
   width: 100px;
   /* display: inline; */
 }
+
+/* .modal-message {
+  display: inline;
+} */
 
 .child {
   padding: 50px;
