--- src/Components/Modal/ModalContent.css
@@ -1,39 +1,0 @@
-/* Modal */
-.backdrop {
-  background-color: rgba(0, 0, 0, 0.3);
-  bottom: 0;
-  left: 0;
-  right: 0;
-  top: 0;
-  position: fixed;
-  display: grid;
-  place-items: center;
-}
-
-.modal-div {
-  background: white;
-  border-radius: 5px;
-  color: rgb(51, 51, 51);
-  display: inline;
-  /* display: flex; */
-  height: 70%;
-  width: 50%;
-  max-height: 620px;
-  max-width: 500px;
-  outline: none;
-  padding: 10px;
-  align-items: center;
-  justify-content: center;
-  transition: opacity 0.2s ease-in-out;
-  z-index: 999;
-  overflow: scroll;
-}
-
-.modal-button-close {
-  width: 100px;
-  /* display: inline; */
-}
-
-/* .modal-message {
-  display: inline;
-} */
