--- src/Components/ModalCambiarPassword/ModalCambiarPassword.css
@@ -1,13 +1,0 @@
-
-.button-change-pss {
-    background-color: #fbfef9;
-    border: solid 2px #fe8b86;
-    color: #ff7a75;
-    width: 382px;
-  }
-  
-  .button-change-pss:hover {
-    background-color: #fe8b86;
-    border: none;
-    color: #fbfef9;
-  }
